pyproject.toml
[project]
name = "simple-weather-app"
version = "0.1.0"
description = "Simple weather app that uses API calls to get data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
